requirements.txt
numpy
opencv-python-headless
Pillow
discord.py
python-dotenv
attrs